feat: フィルターパネルが表示できるようにする
