+++ src/components/pages/Home/index.module.scss
@@ -1,6 +1,7 @@
 @import '../../../assets/styles/resources';
 
 .home-component {
+  position: relative;
   width: 100%;
 
   > .dish-list {
@@ -31,9 +32,17 @@
     }
   }
 
-  > .filter {
+  > .icon {
     position: fixed;
     right: 5%;
     bottom: 3%;
+  }
+
+  > .filter {
+    position: fixed;
+    top: 0;
+    z-index: 1;
+    width: 100%;
+    height: 100%;
   }
 }
