+++ src/components/ui/filter/Filter/index.module.scss
@@ -1,6 +1,7 @@
 @import '../../../../assets/styles/resources';
 
 .filter-component {
+  position: relative;
   width: 100%;
   height: 100%;
 
@@ -19,5 +20,13 @@
       height: 24px;
       color: $color-white;
     }
+  }
+
+  > .panel {
+    position: fixed;
+    top: 0;
+    z-index: 10;
+    width: 100%;
+    height: 100%;
   }
 }
