+++ src/components/ui/filter/FilterPanel/index.module.scss
@@ -1,9 +1,6 @@
 @import '../../../../assets/styles/resources';
 
 .filter-panel-component {
-  position: fixed;
-  top: 0;
-  z-index: 1;
   width: 100%;
   height: 100%;
   padding-top: 56px;
